+++ cvat-ui/src/reducers/interfaces.ts
@@ -117,6 +117,14 @@ export interface ExportState {
     modalVisible: boolean;
 }
 
+export interface ImportState {
+    importingId: number | null;
+    progress: number;
+    status: string;
+    instance: any;
+    modalVisible: boolean;
+}
+
 export interface FormatsState {
     annotationFormats: any;
     fetching: boolean;
@@ -395,6 +403,14 @@ export interface NotificationsState {
         };
         predictor: {
             prediction: null | ErrorState;
+        };
+        exporting: {
+            dataset: null | ErrorState;
+            annotation: null | ErrorState;
+        };
+        importing: {
+            dataset: null | ErrorState;
+            annotation: null | ErrorState;
         };
         cloudStorages: {
             creating: null | ErrorState;
@@ -705,6 +721,7 @@ export interface CombinedState {
     shortcuts: ShortcutsState;
     review: ReviewState;
     export: ExportState;
+    import: ImportState;
     cloudStorages: CloudStoragesState;
 }